:sparkles: ParticipationEntity creation

+++ Sources/src/main/kotlin/allin/entities/UserEntity.kt
@@ -14,13 +14,13 @@ interface UserEntity : Entity<UserEntity> {
     val username: String
     var email: String
     var password: String
-    var nbCoins: Double
+    var nbCoins: Int
 }
 object UsersEntity : Table<UserEntity>("utilisateur") {
     val id = uuid("id").primaryKey()
     val username = varchar("username")
     val password = varchar("password")
-    val nbCoins = double("coins")
+    val nbCoins = int("coins")
     val email = varchar("email")
 
     fun getUserToUserDTO(): MutableList<UserDTO> {
@@ -29,7 +29,7 @@ object UsersEntity : Table<UserEntity>("utilisateur") {
                 row[id].toString(),
                 row[username].toString(),
                 row[email].toString(),
-                row[nbCoins]?:0.0,
+                row[nbCoins]?:0,
                 null
             )
         }.toMutableList()
@@ -51,7 +51,7 @@ object UsersEntity : Table<UserEntity>("utilisateur") {
                         row[id].toString(),
                         row[username].toString(),
                         row[email].toString(),
-                        row[nbCoins] ?: 0.0,
+                        row[nbCoins]?:0,
                         null
                     ),
                     row[password].toString()
